requirements.test.txt
-r requirements.txt
pytest==7.4.3
pytest-django==4.6.0